move rename *-checker.ts to *-check.ts and move them into subfolder
